add different files to separate old_main.py add geometry_classes.py and euler_scenes.py
